third_party/blink/web_tests/external/wpt/storage-access-api/storage-access-beyond-cookies.unpartitioned.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies when unpartitioned
  assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for unpartitioned" but got "Unable to load handle in same-origin context for unpartitioned"
Harness: the test ran to completion.
